流程模型分页查找、分页参数优化

+++ sherly-sdk/src/main/java/com/guzi/sherly/model/PageQuery.java
@@ -12,9 +12,9 @@ import lombok.Data;
 public class PageQuery {
     /** 当前页 */
     @ApiModelProperty(value = "当前页")
-    private Long current = 0L;
+    private Integer current = 0;
 
     /** 页大小 */
     @ApiModelProperty(value = "页大小")
-    private Long size = 10L;
+    private Integer size = 10;
 }
